Search middleware: Ripgrep as service
