+++ crates/middleware/src/lib.rs
@@ -1,6 +1,17 @@
 use std::time;
 use serde::Deserialize;
 use serde_json as json;
+use std::collections::hash_map::DefaultHasher;
+use std::hash::{Hash, Hasher};
+use std::collections::HashSet;
+use std::fs::{self, File};
+use terraphim_types::{Article, ConfigState};
+use std::process::{ExitStatus, Stdio};
+use tokio::process::{Child, Command};
+use tokio::io::AsyncBufRead;
+
+use tokio::io::{AsyncBufReadExt, AsyncReadExt, AsyncWriteExt, BufReader as TokioBufferedReader};
+
 #[derive(Clone, Debug, Deserialize, PartialEq, Eq)]
 #[serde(tag = "type", content = "data")]
 #[serde(rename_all = "snake_case")]
@@ -137,8 +148,162 @@ pub fn json_decode(jsonlines: &str) -> Vec<Message> {
         .collect::<Result<Vec<Message>, _>>()
         .unwrap()
 }
+fn calculate_hash<T: Hash>(t: &T) -> String {
+    let mut s = DefaultHasher::new();
+    t.hash(&mut s);
+    format!("{:x}", s.finish())
+}
+
+pub struct RipgrepService {
+    command: String,
+    args: Vec<String>,
+}
+impl RipgrepService {
+    pub fn new(command: String, args: Vec<String>) -> Self {
+        Self {
+            command,
+            args,
+        }
+    }
+    pub async fn run(&self) -> Result<Vec<Message>, std::io::Error> {
+        let mut child = Command::new(&self.command)
+        .args(&self.args)
+        .stdout(Stdio::piped())
+        .spawn()
+        .unwrap();
+        let mut stdout = child.stdout.take().expect("Stdout is not available"); 
+        let read = async move { 
+            let mut data = String::new(); 
+            stdout.read_to_string(&mut data).await.map(|_| data) 
+        }; 
+        let output = read.await;
+        let msgs = json_decode(&output.unwrap());
+        Ok(msgs)
+    }
+}
+
+/// Service to run and index output of ripgrep into TerraphimGraph
+pub async fn run_ripgrep_service_and_index(mut config_state:ConfigState, needle: String, haystack: String) {
 
 
+    let ripgrep_service = RipgrepService::new("/home/alex/.cargo/bin/rg".to_string(), vec![needle.clone(), haystack.clone(), "--json".to_string(), "--trim".to_string(), "-C3".to_string(), "-i".to_string()]);
+    let msgs = ripgrep_service.run().await.unwrap();
+    
+    let mut article = Article::default();
+
+    let mut existing_paths: HashSet<String> = HashSet::new();
+
+    for each_msg in msgs.iter() {
+        match each_msg {
+            Message::Begin(begin_msg) => {
+                println!("stdout: {:#?}", each_msg);
+                article = Article::default();
+
+                // get path
+                let path: Option<Data> = begin_msg.path.clone();
+                let path_text = match path {
+                    Some(Data::Text{text}) => text,
+                    _ => {
+                        println!("Error: path is not text");
+                        continue
+                    }
+                };
+
+                if existing_paths.contains(&path_text) {
+                    continue
+                }
+                existing_paths.insert(path_text.clone());
+
+                let id = calculate_hash(&path_text);
+                article.id = Some(id.clone());
+                article.title = path_text.clone();
+                article.url = path_text;
+
+            },
+            Message::Match(match_msg) => {
+                println!("stdout: {:#?}", article);
+                let path = match_msg.path.clone();
+                let path = path.unwrap();
+                let path_text = match path {
+                    Data::Text{text} => text,
+                    _ => {
+                        println!("Error: path is not text");
+                        continue
+                    }
+                };
+                let body = fs::read_to_string(path_text).unwrap();
+                article.body = body;
+
+                let lines = match &match_msg.lines {
+                    Data::Text{text} => text,
+                    _ => {
+                        println!("Error: lines is not text");
+                        continue
+                    }
+                };
+                match article.description {
+                    Some(description) => {
+                        article.description = Some(description + " " + &lines);
+                    },
+                    None => {
+                        article.description = Some(lines.clone());
+                    }
+                }
+            },
+            Message::Context(context_msg) => {
+                // let article = Article::new(context_msg.clone());
+                println!("stdout: {:#?}", article);
+
+                let article_url = article.url.clone();
+                let path = context_msg.path.clone();
+                let path = path.unwrap();
+                let path_text = match path {
+                    Data::Text{text} => text,
+                    _ => {
+                        println!("Error: path is not text");
+                        continue
+                    }
+                };
+
+                // We got a context for a different article
+                if article_url != path_text {
+                    continue
+                }
+
+                let lines = match &context_msg.lines {
+                    Data::Text{text} => text,
+                    _ => {
+                        println!("Error: lines is not text");
+                        continue
+                    }
+                };
+                match article.description {
+                    Some(description) => {
+                        article.description = Some(description + " " + &lines);
+                    },
+                    None => {
+                        article.description = Some(lines.clone());
+                    }
+                }
+            },
+            Message::End(end_msg) => {
+                println!("stdout: {:#?}", each_msg);
+                // The `End` message could be received before the `Begin` message
+                // causing the article to be empty
+                let id = match article.id {
+                    Some(ref id) => id,
+                    None => {
+                        continue
+                    }
+                };
+                config_state.index_article(article.clone()).await.expect("Failed to index article");
+            },
+            _ => {
+            }
+        }
+        
+    }
+}
 #[cfg(test)]
 mod tests {
     use super::*;
